+++ src/components/TemplatedCaseStudy/FirstTextSection.js
@@ -1,0 +1,24 @@
+import React from 'react'
+import styled from 'styled-components'
+import breakpoint from 'styled-components-breakpoint'
+
+import { BodyPrimary } from '../Typography'
+import { Row, Col } from '../grid'
+
+const RowLayout = styled(Row)`
+  ${breakpoint('smallTablet')`
+    justify-content: flex-end;
+  `}
+`
+
+const FirstTextSection = ({ textAboveVideo }) => (
+  <RowLayout>
+    <Col width={[1, 1, 1, 9 / 12, 7 / 12]}>
+      {textAboveVideo.map((text, i) => (
+        <BodyPrimary key={i}>{text}</BodyPrimary>
+      ))}
+    </Col>
+  </RowLayout>
+)
+
+export default FirstTextSection
